% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_data}
\alias{grid_data}
\title{Construct grids over time, longitude, latitude, and all combinations thereof}
\usage{
grid_data(x, longitude, latitude, time)
}
\description{
Construct grids over time, longitude, latitude, and all combinations thereof
}
