% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{read_nc_grid_data}
\alias{read_nc_grid_data}
\title{Read NetCDF file to a grid format}
\usage{
read_nc_grid_data(
  filename,
  variable_name,
  longitude_name = "lon",
  latitude_name = "lat",
  time_name = "time"
)
}
\description{
Read NetCDF file to a grid format
}
