% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{write_nc_grid_data}
\alias{write_nc_grid_data}
\title{Write grid to NetCDF file}
\usage{
write_nc_grid_data(
  x,
  filename,
  variable_name = "value",
  variable_unit = "1",
  variable_longname = "Scaling factor"
)
}
\description{
Write grid to NetCDF file
}
